requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
sqlparse==0.5.1
typing_extensions==4.12.2
numpy==2.0.1
matplotlib==3.9.1.post1
scikit-fuzzy==0.4.2
